Json Communication
Rather than using serialized java objects, communication is now done by serializing to JSON and back.

This requires keeping a registry of HexMessage objects to use for conversion. Pending a formal registry system this is mocked into both the client and server.

JSON serialization is performed using the GSON library.